requirements.txt
gradio>=5.35.0
openai>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.31.0
numpy>=1.26.4
pypdf>=4.2.0
fastapi>=0.115.0
uvicorn>=0.30.0

python-multipart>=0.0.9
